Add comments for thread handler for tickSystem

+++ src/main/scala/com/rpg/game/tick/TimeTickSystem.scala
@@ -5,21 +5,21 @@ import com.badlogic.gdx.{ApplicationAdapter, Gdx, Screen, ScreenAdapter}
 import scala.collection.mutable
 
 class TimeTickSystem extends ApplicationAdapter {
-  
+
   private val TICK_TIMER_MAX: Float = .2f //every .2f == 200 milliseconds or 5 ticks per second
-  private var tick: Int = 0
+  private var tick: Long = 0
   private var tickTimer: Float = 0.0f
-  
-  
+
+
   //list of event handlers
-  private val tickEventHandlers: mutable.ListBuffer[Int => Unit] = mutable.ListBuffer()
-  
-  def getCurrentTick: Int = tick
-  
-  def addTickEventHandler(handler: Int => Unit): Unit = {
+  private val tickEventHandlers: mutable.ListBuffer[Long => Unit] = mutable.ListBuffer()
+
+  def getCurrentTick: Long = tick
+
+  def addTickEventHandler(handler: Long => Unit): Unit = {
     tickEventHandlers += handler
   }
-  
+
   override def create(): Unit = {
     tick = 0
   }
@@ -27,14 +27,15 @@ class TimeTickSystem extends ApplicationAdapter {
 
   override def render(): Unit = {
     tickTimer += Gdx.graphics.getDeltaTime
-
-    if(tickTimer >= TICK_TIMER_MAX){ 
+    if(tickTimer >= TICK_TIMER_MAX){
       tickTimer -= TICK_TIMER_MAX
       tick += 1
-      
+
+      //3 threads to play with for handler, 1 for display, 1 tickSystem (spins off 3 other threads to handle results) 
+      //blocking so multi-thread
       if(tick % 5 == 0) tickEventHandlers.foreach(handler => handler(tick))
     }
   }
-  
-  
+
+
 }
